# Speedup logging